Fake input and bug fixes
 * Press W in simulator to activate (pretty slow shit, do avoid unless compiling with optimisations)
 * Tiny improvement of the EEPROM simulation.
 * Fixed type typo for millis and micros

+++ simulation/include/Arduino.h
@@ -73,8 +73,8 @@ extern SimSerial Serial;
 
 //extern void putString(int row, int col, int color, const char* msg, const FONT_INFO fontInfo);
 
-uint16_t micros();
-uint16_t millis();
+uint32_t micros();
+uint32_t millis();
 void delay(uint32_t millis);
 void delayMicroseconds(uint32_t micros);
